pyproject.toml
[tool.poetry]
name = "funexpression"
version = "0.1.0"
description = ""
authors = ["Adeonita <adeonita.sousa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
bio = "^1.7.1"
python-dotenv = "^1.0.1"
celery = {extras = ["rabbitmq"], version = "^5.4.0"}
pysondb = "^1.6.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
